TestMain を main_test.go に移動する
